Moved CRUDL methods to client.go
